requirements.txt
pdfplumber
pdf2image
PyMuPDF
Pillow
opencv-python
numpy
pandas
concurrent.futures
requests
pytesseract 